+++ curvefs/src/volume/block_device_client.cpp
@@ -22,6 +22,7 @@
 
 #include "curvefs/src/volume/block_device_client.h"
 
+#include <sys/types.h>
 #include <glog/logging.h>
 
 #include <map>
@@ -52,10 +53,10 @@ BlockDeviceClientImpl::BlockDeviceClientImpl()
     : fd_(-1), fileClient_(std::make_shared<FileClient>()) {}
 
 BlockDeviceClientImpl::BlockDeviceClientImpl(
-    const std::shared_ptr<FileClient>& fileClient)
+    const std::shared_ptr<FileClient> &fileClient)
     : fd_(-1), fileClient_(fileClient) {}
 
-bool BlockDeviceClientImpl::Init(const BlockDeviceClientOptions& options) {
+bool BlockDeviceClientImpl::Init(const BlockDeviceClientOptions &options) {
     auto ret = fileClient_->Init(options.configPath);
     if (ret != LIBCURVE_ERROR::OK) {
         LOG(ERROR) << "Init file client error: " << ret;
@@ -65,12 +66,10 @@ bool BlockDeviceClientImpl::Init(const BlockDeviceClientOptions& options) {
     return true;
 }
 
-void BlockDeviceClientImpl::UnInit() {
-    fileClient_->UnInit();
-}
+void BlockDeviceClientImpl::UnInit() { fileClient_->UnInit(); }
 
-bool BlockDeviceClientImpl::Open(const std::string& filename,
-                                 const std::string& owner) {
+bool BlockDeviceClientImpl::Open(const std::string &filename,
+                                 const std::string &owner) {
     UserInfo userInfo(owner);
     curve::client::OpenFlags flags;
     auto retCode = fileClient_->Open(filename, userInfo, flags);
@@ -101,9 +100,9 @@ bool BlockDeviceClientImpl::Close() {
     return true;
 }
 
-bool BlockDeviceClientImpl::Stat(const std::string& filename,
-                                          const std::string& owner,
-                                          BlockDeviceStat* statInfo) {
+bool BlockDeviceClientImpl::Stat(const std::string &filename,
+                                 const std::string &owner,
+                                 BlockDeviceStat *statInfo) {
     FileStatInfo fileStatInfo;
     UserInfo userInfo(owner);
     auto retCode = fileClient_->StatFile(filename, userInfo, &fileStatInfo);
@@ -122,7 +121,7 @@ bool BlockDeviceClientImpl::Stat(const std::string& filename,
     return true;
 }
 
-ssize_t BlockDeviceClientImpl::Read(char* buf, off_t offset, size_t length) {
+ssize_t BlockDeviceClientImpl::Read(char *buf, off_t offset, size_t length) {
     VLOG(9) << "read request, offset: " << offset << ", length: " << length;
 
     LatencyUpdater updater(&g_read_latency);
@@ -138,7 +137,7 @@ ssize_t BlockDeviceClientImpl::Read(char* buf, off_t offset, size_t length) {
     return request.Wait();
 }
 
-ssize_t BlockDeviceClientImpl::Readv(const std::vector<ReadPart>& iov) {
+ssize_t BlockDeviceClientImpl::Readv(const std::vector<ReadPart> &iov) {
     if (iov.size() == 1) {
         VLOG(9) << "read block offset: " << iov[0].offset
                 << ", length: " << iov[0].length;
@@ -148,7 +147,7 @@ ssize_t BlockDeviceClientImpl::Readv(const std::vector<ReadPart>& iov) {
     std::vector<std::unique_ptr<AioRead>> requests;
     requests.reserve(iov.size());
 
-    for (const auto& io : iov) {
+    for (const auto &io : iov) {
         requests.push_back(absl::make_unique<AioRead>(
             io.offset, io.length, io.data, fileClient_.get(), fd_));
 
@@ -157,7 +156,7 @@ ssize_t BlockDeviceClientImpl::Readv(const std::vector<ReadPart>& iov) {
 
     bool error = false;
     ssize_t total = 0;
-    for (const auto& r : requests) {
+    for (const auto &r : requests) {
         auto nr = r->Wait();
         if (nr < 0) {
             error = true;
@@ -171,8 +170,7 @@ ssize_t BlockDeviceClientImpl::Readv(const std::vector<ReadPart>& iov) {
     return error ? -1 : total;
 }
 
-ssize_t BlockDeviceClientImpl::Write(const char* buf,
-                                     off_t offset,
+ssize_t BlockDeviceClientImpl::Write(const char *buf, off_t offset,
                                      size_t length) {
     VLOG(9) << "write request, offset: " << offset << ", length: " << length;
 
@@ -189,7 +187,7 @@ ssize_t BlockDeviceClientImpl::Write(const char* buf,
     return request.Wait();
 }
 
-ssize_t BlockDeviceClientImpl::Writev(const std::vector<WritePart>& iov) {
+ssize_t BlockDeviceClientImpl::Writev(const std::vector<WritePart> &iov) {
     if (iov.size() == 1) {
         return Write(iov[0].data, iov[0].offset, iov[0].length);
     }
@@ -197,7 +195,7 @@ ssize_t BlockDeviceClientImpl::Writev(const std::vector<WritePart>& iov) {
     std::vector<std::unique_ptr<AioWrite>> requests;
     requests.reserve(iov.size());
 
-    for (const auto& io : iov) {
+    for (const auto &io : iov) {
         requests.push_back(absl::make_unique<AioWrite>(
             io.offset, io.length, io.data, fileClient_.get(), fd_));
 
@@ -206,7 +204,7 @@ ssize_t BlockDeviceClientImpl::Writev(const std::vector<WritePart>& iov) {
 
     bool error = false;
     ssize_t total = 0;
-    for (const auto& r : requests) {
+    for (const auto &r : requests) {
         auto nr = r->Wait();
         if (nr < 0) {
             error = true;
@@ -220,12 +218,11 @@ ssize_t BlockDeviceClientImpl::Writev(const std::vector<WritePart>& iov) {
     return error ? -1 : total;
 }
 
-bool BlockDeviceClientImpl::WritePadding(char* writeBuffer,
-                                         off_t writeStart,
+bool BlockDeviceClientImpl::WritePadding(char *writeBuffer, off_t writeStart,
                                          off_t writeEnd,
                                          off_t offset,     // actual offset
                                          size_t length) {  // actual length
-    std::vector<std::pair<off_t, size_t>> readvec;  // Align reads
+    std::vector<std::pair<off_t, size_t>> readvec;         // Align reads
     off_t readEnd = 0;
 
     // Padding leading
@@ -235,7 +232,8 @@ bool BlockDeviceClientImpl::WritePadding(char* writeBuffer,
     }
 
     // Padding trailing
-    if (offset + length > readEnd && offset + length != writeEnd) {
+    if (static_cast<off_t>(offset + length) > readEnd &&
+        static_cast<off_t>(offset + length) != writeEnd) {
         off_t readStart = writeEnd - IO_ALIGNED_BLOCK_SIZE;
         if (readvec.size() == 1 && readStart == readEnd) {
             readvec[0].second = IO_ALIGNED_BLOCK_SIZE * 2;
@@ -244,10 +242,10 @@ bool BlockDeviceClientImpl::WritePadding(char* writeBuffer,
         }
     }
 
-    for (const auto& item : readvec) {
+    for (const auto &item : readvec) {
         auto retCode = AlignRead(writeBuffer + item.first - writeStart,
                                  item.first, item.second);
-        if (retCode != item.second) {
+        if (retCode != static_cast<ssize_t>(item.second)) {
             return false;
         }
     }
@@ -255,14 +253,13 @@ bool BlockDeviceClientImpl::WritePadding(char* writeBuffer,
     return true;
 }
 
-ssize_t BlockDeviceClientImpl::AlignRead(char* buf,
-                                         off_t offset,
+ssize_t BlockDeviceClientImpl::AlignRead(char *buf, off_t offset,
                                          size_t length) {
     auto ret = fileClient_->Read(fd_, buf, offset, length);
     if (ret < 0) {
         LOG(ERROR) << "Read file failed, retCode = " << ret;
         return -1;
-    } else if (ret != length) {
+    } else if (static_cast<size_t>(ret) != length) {
         LOG(ERROR) << "Read file failed, expect read " << length
                    << " bytes, actual read " << ret << " bytes";
         return -1;
@@ -271,14 +268,13 @@ ssize_t BlockDeviceClientImpl::AlignRead(char* buf,
     return length;
 }
 
-ssize_t BlockDeviceClientImpl::AlignWrite(const char* buf,
-                                          off_t offset,
+ssize_t BlockDeviceClientImpl::AlignWrite(const char *buf, off_t offset,
                                           size_t length) {
     auto ret = fileClient_->Write(fd_, buf, offset, length);
     if (ret < 0) {
         LOG(ERROR) << "Write file failed, retCode = " << ret;
         return -1;
-    } else if (ret != length) {
+    } else if (static_cast<size_t>(ret) != length) {
         LOG(ERROR) << "Write file failed, expect write " << length
                    << " bytes, actual write " << ret << " bytes";
         return -1;
@@ -288,15 +284,14 @@ ssize_t BlockDeviceClientImpl::AlignWrite(const char* buf,
 }
 
 bool BlockDeviceClientImpl::ConvertFileStatus(int fileStatus,
-                                              BlockDeviceStatus* bdStatus) {
-    static const std::map<int, BlockDeviceStatus> fileStatusMap {
-        { 0, BlockDeviceStatus::CREATED },
-        { 1, BlockDeviceStatus::DELETING },
-        { 2, BlockDeviceStatus::CLONING },
-        { 3, BlockDeviceStatus::CLONE_META_INSTALLED },
-        { 4, BlockDeviceStatus::CLONED },
-        { 5, BlockDeviceStatus::BEING_CLONED }
-    };
+                                              BlockDeviceStatus *bdStatus) {
+    static const std::map<int, BlockDeviceStatus> fileStatusMap{
+        {0, BlockDeviceStatus::CREATED},
+        {1, BlockDeviceStatus::DELETING},
+        {2, BlockDeviceStatus::CLONING},
+        {3, BlockDeviceStatus::CLONE_META_INSTALLED},
+        {4, BlockDeviceStatus::CLONED},
+        {5, BlockDeviceStatus::BEING_CLONED}};
 
     auto iter = fileStatusMap.find(fileStatus);
     if (iter == fileStatusMap.end()) {
